backend/requirements.txt
# VoiceAI Backend Dependencies
pipecat-ai>=0.0.40
openai>=1.0.0
aiohttp>=3.9.0
websockets>=12.0
python-dotenv>=1.0.0
loguru>=0.7.0
